+++ changelog.md
@@ -4,6 +4,7 @@
 
 **New Features**
 * Path to AAS Environment model can be set in persistence config
+* Add basic for basic username & password authentication for asset connections (MQTT, OPC UA, HTTP)
 
 **Internal changes & Bugfixes**
 *